Separate user & session tests
